% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightbox.R
\name{lightbox_ui}
\alias{lightbox_ui}
\alias{lightbox}
\alias{use_lightbox}
\title{Lightbox Image Gallery}
\usage{
lightbox_ui(id)

lightbox(input, output, session, src, gallery, caption = "",
  display = "block")

use_lightbox()
}
\arguments{
\item{src}{A vector of paths to images.}

\item{gallery}{An identifier for an image gallery.}

\item{caption}{A vector of same length as \code{src} with corresponding image captions.}

\item{display}{CSS display property.}
}
\description{
A shiny module to be used inside a shiny app. Make sure to include
\code{\link[=use_lightbox]{use_lightbox()}} in the UI to load the necessary javascript and css ressources.
}
